이벤트 미션 장소 수정 안되도록

+++ app/Http/Controllers/v1_1/FeedController.php
@@ -528,7 +528,6 @@ class FeedController extends Controller
                 $feed->update($update_data);
             }
 
-            if (count($feed->missions()->where('is_event', 1)) == 0) {
                 if ($product_delete) {
                     $feed->product()->delete();
                 } elseif ($product_id) {
@@ -546,21 +545,21 @@ class FeedController extends Controller
                     ]);
                     $feed->product()->updateOrCreate([], ['type' => 'outside', 'outside_product_id' => $product->id]);
                 }
-            }
 
-
-            if ($place_delete) {
-                $feed->feed_place()->delete();
-            } elseif ($place_address && $place_title) {
-                $place = Place::updateOrCreate(['title' => $place_title], [
-                    'address' => $place_address,
-                    'description' => $place_description,
-                    'image' => $place_image,
-                    'url' => $place_url ?? urlencode("https://google.com/search?q=$place_title"),
-                    'lat' => $place_lat,
-                    'lng' => $place_lng,
-                ]);
-                $feed->feed_place()->updateOrCreate([], ['place_id' => $place->id]);
+            if (count($feed->missions()->where('is_event', 1)) == 0) {
+                if ($place_delete) {
+                    $feed->feed_place()->delete();
+                } elseif ($place_address && $place_title) {
+                    $place = Place::updateOrCreate(['title' => $place_title], [
+                        'address' => $place_address,
+                        'description' => $place_description,
+                        'image' => $place_image,
+                        'url' => $place_url ?? urlencode("https://google.com/search?q=$place_title"),
+                        'lat' => $place_lat,
+                        'lng' => $place_lng,
+                    ]);
+                    $feed->feed_place()->updateOrCreate([], ['place_id' => $place->id]);
+                }
             }
 
             DB::commit();
